Update volume type from double to int.

+++ src/main/java/com/drinkwater/apidrinkwater/hydrationtracking/dto/WaterIntakeCreateDTO.java
@@ -13,7 +13,7 @@ public class WaterIntakeCreateDTO {
     private Long userId;
 
     @Positive(message = "Volume must be a positive number.")
-    private double volume;
+    private int volume;
 
     @NotNull(message = "Date and time of intake is required.")
     @PastOrPresent(message = "Date and time of intake must be in the past or present.")
