+++ src/main/java/com/drinkwater/apidrinkwater/hydrationtracking/dto/WaterIntakeResponseDTO.java
@@ -10,7 +10,7 @@ public class WaterIntakeResponseDTO {
 
     private Long id;
     private OffsetDateTime dateTimeUTC;
-    private double volume;
+    private int volume;
     private VolumeUnit volumeUnit;
     private Long userId;
 }
